+++ server/apps/packages/__init__.py
@@ -1,0 +1,2 @@
+from .package_service import PackageService  # noqa
+from .takes_package_service import TakesPackageService  # noqa
